+++ PyInstaller/hooks/hook-IPython.py
@@ -8,10 +8,20 @@
 #-----------------------------------------------------------------------------
 
 
-from PyInstaller.utils.hooks import (collect_data_files, collect_submodules)
+# Tested with IPython 4.0.0.
+
+from PyInstaller.compat import modname_tkinter, is_win, is_darwin
+from PyInstaller.utils.hooks import collect_data_files, collect_submodules
+
+# Ignore 'matplotlib'. IPython contains support for matplotlib.
+# Ignore GUI libraries. IPython supports integration with GUI frameworks.
+# Assume that it will be imported by any other module when the user really
+# uses it.
+excludedimports = ['gtk', 'matplotlib', 'PyQt4', 'PyQt5', 'PySide']
+
+# IPython uses 'tkinter' for clipboard access on Linux/Unix. Exclude it on Windows and OS X.
+if is_win or is_darwin:
+    excludedimports.append(modname_tkinter)
 
 
-# IPython (tested with 0.13) requires the following files:
-#   ./site-packages/IPython/config/profile/README_STARTUP
 datas = collect_data_files('IPython')
-hiddenimports = collect_submodules('IPython')