+++ PyInstaller/hooks/hook-sqlalchemy.py
@@ -10,6 +10,9 @@
 
 from PyInstaller.utils.hooks import exec_statement, is_module_satisfies
 
+# 'sqlalchemy.testing' causes bundling a lot of unnecessary modules.
+excludedimports = ['sqlalchemy.testing']
+
 # include most common database bindings
 # some database bindings are detected and include some
 # are not. We should explicitly include database backends.