+++ tests/functional/test_import.py
@@ -116,10 +116,10 @@ def test_import_respects_path(pyi_builder, script_dir):
       ['--additional-hooks-dir='+script_dir.join('pyi_hooks').strpath])
 
 
-def test_import_pyqt5_uic_port(pyi_builder):
+def test_import_pyqt5_uic_port(monkeypatch, pyi_builder):
     extra_path = os.path.join(_MODULES_DIR, 'pyi_import_pyqt_uic_port')
     pyi_builder.test_script('pyi_import_pyqt5_uic_port.py',
-                            pyi_args=['--path', extra_path], )
+                            pyi_args=['--path', extra_path])
 
 
 #--- ctypes ----