+++ doc/CHANGES.rst
@@ -4,7 +4,7 @@ Changelog for PyInstaller
 3.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- New Hooks: traitlets,
 
 
 3.1.1 (2016-01-31)